requirements.txt
# =========================================================================
# REQUIREMENTS.TXT - GOALMANAGER BACKEND
# =========================================================================
# Dependências Python para API FastAPI com MySQL
# =========================================================================

# FastAPI e servidor
fastapi==0.115.4
uvicorn[standard]==0.32.0

# Banco de dados
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
alembic==1.13.3

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validação e serialização
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator==2.2.0

# Utilitários
python-dotenv==1.0.0
python-dateutil==2.8.2

# Desenvolvimento e testes
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Logs e monitoramento
loguru==0.7.2

# Geração de documentação
markdown==3.5.1

# Validação de dados adicionais
validators==0.22.0

# Timezone
pytz==2023.3